refactor(#5): split bulk import dialog into hook and panels
Extract types, dialog state/import pipeline, dropzone, groups panel, and footer; keep BulkImportGroupItem re-exported from the dialog entry.
